% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assetservers.R
\name{pi_assetserver_list}
\alias{pi_assetserver_list}
\alias{pi_assetserver}
\title{Retrieve a list of all Asset Servers known to this service}
\usage{
pi_assetserver_list(.con, ...)

pi_assetserver(.con, webId, ...)
}
\arguments{
\item{.con}{A \code{\link[=pi_connect]{pi_connect()}} object}

\item{...}{Passed to \code{\link[=pi_get]{pi_get()}}}

\item{webId}{The ID of the server.}
}
\description{
Retrieve a list of all Asset Servers known to this service
}
\examples{
con <- pi_connect_public()
pi_assetserver_list(con)
pi_assetserver(con, "F1RSIRAQC7zjPUOfBqai218IAwUElTUlYx")

}
\references{
https://devdata.osisoft.com/piwebapi/help/controllers/assetserver/actions/list
https://devdata.osisoft.com/piwebapi/help/controllers/assetserver/actions/get
}
