% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{pi_get}
\alias{pi_get}
\alias{pi_response}
\alias{pi_url}
\title{Low-level API access}
\usage{
pi_get(.con, .fun = NULL, ..., .quiet = TRUE)

pi_response(.con, .fun = NULL, ..., .verb = "GET", .quiet = TRUE)

pi_url(.con, .fun = NULL, ...)
}
\arguments{
\item{.con}{A \code{\link[=pi_connect]{pi_connect()}} object}

\item{.fun}{A function ame}

\item{...}{Key-value query string parameters. This uses \code{\link[rlang:list2]{rlang::list2()}},
so splicing (i.e., \code{!!!}) is supported.}

\item{.quiet}{Use \code{.quiet = FALSE} to see the URL calls}

\item{.verb}{The HTTP verb to use (see \code{\link[httr:VERB]{httr::VERB()}}).}
}
\value{
\code{\link[=pi_get]{pi_get()}} returns a \code{\link[=list]{list()}}, \code{\link[=pi_response]{pi_response()}} returns an
\code{\link[httr:response]{httr::response()}}, and \code{\link[=pi_url]{pi_url()}} returns a bare URL.
}
\description{
Low-level API access
}
\examples{
con <- pi_connect_public()
pi_url(con, "assetservers")
pi_response(con, "assetservers")
result <- pi_get(con, "assetservers")
result$Items[[1]]$Name

}
