% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{pi_web_id}
\alias{pi_web_id}
\title{Extract a Web ID}
\usage{
pi_web_id(webId)
}
\arguments{
\item{webId}{The ID of the object or the object itself.}
}
\value{
A character vector of web IDs
}
\description{
Extract a Web ID
}
