% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{pi_connect}
\alias{pi_connect}
\alias{pi_autheticate}
\title{Connect to the PI web API}
\usage{
pi_connect(endpoint, user = Sys.getenv("R_PI_USER"))

pi_autheticate(user = Sys.getenv("R_PI_USER"))
}
\arguments{
\item{endpoint}{The base URL for the endpoint.}

\item{user}{Username for authentication.}
}
\value{
A connection object
}
\description{
Connects to a web API endpoint, optionally with authentication. \emph{Never store
passwords in script files or type them into the console!!!!} Instead, this
function will propmpt you for a password in RStudio, or you can set the
R_PI_PASSWORD environment variable (you could do so using
\code{\link[usethis:edit_r_environ]{usethis::edit_r_environ()}}).
}
